Fixed typo.

+++ src/modules/Image/encodings/psd.c
@@ -116,11 +116,11 @@ static int read_char( struct buffer *from )
 }
 
 
-static char *read_data( struct buffer * from, unsigned int len )
+static char *read_data( struct buffer * from, size_t len )
 {
   char *res;
   if( from->len < len )
-    error("Not enough space for %u bytes\n", len);
+    error("Not enough space for %lu bytes\n", len);
   res = (char *)from->str;
   from->str += len;
   from->len -= len;
